extract more common cdc logic
